+++ CHANGES.rst
@@ -3,6 +3,8 @@ iiif changelog
 
 2015-XX-XX v0.6.1
 - Now works with python 2.6, 2.7, 3.3 and 3.4
+- Static tile generation supports both use of canonical size syntax for
+  OpenSeadragon >= 1.2.1, and the old syntax for earlier versions
 
 2015-08-17 v0.6.0
 - Refactor manipulators for easier testing